Integrate Spring Data JPA and H2 Database
This commit introduces the following changes to the codebase:

- Added dependencies for `Spring Data JPA` and `H2 Database`
- Modified database configuration in `application-dev.properties`
- Added JPA and H2 Console configurations in `application.properties`

Purpose:
The purpose of this commit is to enhance the project by incorporating `Spring Data JPA` for data access and integrating `H2 Database`. The changes in database configuration aim to set up a test database for development, and additional configurations are added for `JPA` and `H2 Console`.